+++ src/main/java/cl/dsoto/incomes/repositories/MonthRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by root on 13-10-22.
  */
-public interface MonthRepository extends JpaRepository<Month, Integer> {
+public interface MonthRepository extends JpaRepository<Month, Long> {
 }
